Extracted class TypeUtils from BeanUtils and related changes in code and test cases
