+++ test/org.jkiss.dbeaver.test.platform/src/org/jkiss/utils/BeanUtilsTest.java
@@ -110,111 +110,6 @@ public class BeanUtilsTest {
     }
 
     @Test
-    public void testIsArrayType() {
-        Assert.assertFalse(BeanUtils.isArrayType(String.class));
-    }
-
-    @Test
-    public void testIsCollectionType() {
-        Assert.assertFalse(BeanUtils.isCollectionType(String.class));
-    }
-
-    @Test
-    public void testGetCollectionType1() {
-        ParameterizedType parameterizedType = new ParameterizedType() {
-            @Override
-            public Type[] getActualTypeArguments() {
-                return new Type[]{String.class};
-            }
-
-            @Override
-            public Type getRawType() {
-                return null;
-            }
-
-            @Override
-            public Type getOwnerType() {
-                return null;
-            }
-        };
-
-        Assert.assertEquals(String.class,
-                BeanUtils.getCollectionType(parameterizedType));
-
-        Assert.assertNull(BeanUtils.getCollectionType(null));
-    }
-
-    @Test
-    public void testGetCollectionType2() {
-        WildcardType wildcardType = new WildcardType() {
-            @Override
-            public Type[] getUpperBounds() {
-                return new Type[]{String.class, Integer.class};
-            }
-
-            @Override
-            public Type[] getLowerBounds() {
-                return new Type[0];
-            }
-        };
-
-        ParameterizedType parameterizedType = new ParameterizedType() {
-            @Override
-            public Type[] getActualTypeArguments() {
-                return new Type[]{wildcardType};
-            }
-
-            @Override
-            public Type getRawType() {
-                return null;
-            }
-
-            @Override
-            public Type getOwnerType() {
-                return null;
-            }
-        };
-
-        Assert.assertEquals(String.class,
-                BeanUtils.getCollectionType(parameterizedType));
-    }
-
-    @Test
-    public void testGetCollectionType3() {
-        WildcardType wildcardType = new WildcardType() {
-            @Override
-            public Type[] getUpperBounds() {
-                return new Type[0];
-            }
-
-            @Override
-            public Type[] getLowerBounds() {
-                return new Type[]{String.class, Integer.class};
-            }
-        };
-
-        ParameterizedType parameterizedType = new ParameterizedType() {
-            @Override
-            public Type[] getActualTypeArguments() {
-                return new Type[]{wildcardType};
-            }
-
-            @Override
-            public Type getRawType() {
-                return null;
-            }
-
-            @Override
-            public Type getOwnerType() {
-                return null;
-            }
-        };
-
-        Assert.assertEquals(String.class,
-                BeanUtils.getCollectionType(parameterizedType));
-    }
-
-    @Test
     public void testReadObjectProperty()
             throws InvocationTargetException, IllegalAccessException {
         Assert.assertNull(BeanUtils.readObjectProperty(null, ".length"));
@@ -223,51 +118,6 @@ public class BeanUtilsTest {
 
         Assert.assertEquals(String.class,
                 BeanUtils.readObjectProperty(String.class, "."));
-    }
-
-    @Test
-    public void testIsBooleanType() {
-        Assert.assertTrue(BeanUtils.isBooleanType(Boolean.TYPE));
-        Assert.assertTrue(BeanUtils.isBooleanType(Boolean.class));
-
-        Assert.assertFalse(BeanUtils.isBooleanType(String.class));
-        Assert.assertFalse(BeanUtils.isBooleanType(null));
-    }
-
-    @Test
-    public void testGetDefaultPrimitiveValue() {
-        Assert.assertEquals(0L,
-                BeanUtils.getDefaultPrimitiveValue(Long.TYPE));
-        Assert.assertEquals(0,
-                BeanUtils.getDefaultPrimitiveValue(Integer.TYPE));
-        Assert.assertEquals(0.0f,
-                BeanUtils.getDefaultPrimitiveValue(Float.TYPE));
-        Assert.assertEquals(0.0,
-                BeanUtils.getDefaultPrimitiveValue(Double.TYPE));
-        Assert.assertEquals((short) 0,
-                BeanUtils.getDefaultPrimitiveValue(Short.TYPE));
-        Assert.assertEquals((byte) 0,
-                BeanUtils.getDefaultPrimitiveValue(Byte.TYPE));
-        Assert.assertEquals((char) 0,
-                BeanUtils.getDefaultPrimitiveValue(Character.TYPE));
-        Assert.assertEquals(false,
-                BeanUtils.getDefaultPrimitiveValue(Boolean.TYPE));
-
-        thrown.expect(IllegalArgumentException.class);
-        BeanUtils.getDefaultPrimitiveValue(String.class);
-    }
-
-    @Test
-    public void testIsNumericType() {
-        Assert.assertTrue(BeanUtils.isNumericType(Long.TYPE));
-        Assert.assertTrue(BeanUtils.isNumericType(Byte.TYPE));
-        Assert.assertTrue(BeanUtils.isNumericType(Short.TYPE));
-        Assert.assertTrue(BeanUtils.isNumericType(Float.TYPE));
-        Assert.assertTrue(BeanUtils.isNumericType(Double.TYPE));
-        Assert.assertTrue(BeanUtils.isNumericType(Short.class));
-        Assert.assertTrue(BeanUtils.isNumericType(Integer.TYPE));
-
-        Assert.assertFalse(BeanUtils.isNumericType(String.class));
     }
 
     @Test
